Fix implicit truncation in CvTacticalUnit::UpdatePriority
m_iAttackPriority was short; UpdatePriority assigns:
  m_iAttackPriority = m_iAttackStrength * m_iHealthPercent

Both short operands are integer-promoted to int before
multiplication. With m_iAttackStrength=1080 and full health=100:
  1080 * 100 = 108000 -> stored in short (max 32767)
  108000 & 0xFFFF = 42464 -> signed short = -23072

UBSan: implicit conversion (sign change), value 108000 -> 42464.

Attack strength scale: GetMaxAttackStrength returns
  GetBaseCombatStrength() * (iModifier + 100)
so a base-10 unit with +8% modifier gives 10*108 = 1080.
The max short*short product is 32767*100 = 3,276,700 -- well
within int range but never within short range at full health.

Bug impact: m_iAttackPriority = -23072 instead of 108000.
operator< sorts descending by priority, so a strong healthy
unit sorts after weaker/damaged ones -- tactical attack order
is wrong for any unit whose strength * health > 32767.

Fix: widen m_iAttackPriority from short to int.
No other short fields are affected: health(0-100),
movesToTarget, expectedDamage values all stay within short.

Log: Shedet (player 43) turn 3, Zone 1150 FRIENDLY,
  posture P_COUNTERATTACK (PlayerTacticalAILog_Shedet.csv),
  targeting enemy combat unit at 79:40 (plot 5199, owner -1)
  via FindUnitsWithinStrikingDistance -> SetHealthPercent
  -> UpdatePriority.

+++ CvGameCoreDLL_Expansion2/CvTacticalAI.h
@@ -130,7 +130,7 @@ private:
 	short m_iMovesToTarget;
 	short m_iExpectedTargetDamage;
 	short m_iExpectedSelfDamage;
-	short m_iAttackPriority;
+	int m_iAttackPriority;
 	CvUnit* m_pInterceptor;
 };
